refactor: Rely on the Instance to notify puppeteers
Previously, we were using a Registry with duplicated keys, but they
are not very error-friendly. In particular, the documentation states:

> For example, if a process crashes, its keys are automatically
> removed from the registry but the change may not propagate
> immediately.

In lkn, we rely on keys rather than PID to identify and reach
processes. The Instance already remembers its registered Puppeteers
key. As a consequence, we can rely on the Instance to dispatch
notification, we do not need the Registry.