feature: Use Instance as proxy for reaching Systems
This patch is a good step towards reducing the risk of race condition,
regarding the insertion of new puppets in a given instance. Indeed,
the `register_puppet` functions of `System` is blocking (it is a
`call`), therefore an Instance will not be able to deal with proxy
requests targeting Systems before the new puppet has been registered
to each Systems.